🚚 Rename mock files
