+++ src/mocks/client/setup.ts
@@ -1,4 +1,4 @@
 import { setupWorker } from 'msw';
-import { handlers } from './handlers.worker';
+import { handlers } from './handlers';
 
 export const worker = setupWorker(...handlers);
