+++ src/mocks/inject.ts
@@ -7,12 +7,12 @@ import { browser, dev } from '$app/environment';
 export async function inject() {
 	if (dev && browser && import.meta.env.VITE_MSW_SERVER_ONLY !== 'true') {
 		console.log('Client MSW started');
-		const { worker } = await import('./client/worker');
+		const { worker } = await import('./client/setup');
 		return worker.start({ onUnhandledRequest: 'bypass' }).catch(console.warn);
 	}
 	if (dev && !browser) {
 		console.log('Server MSW started');
-		const { server } = await import('./server/server');
+		const { server } = await import('./server/setup');
 		return server.listen({ onUnhandledRequest: 'bypass' });
 	}
 }
