+++ src/mocks/server/setup.ts
@@ -1,4 +1,4 @@
 import { setupServer } from 'msw/node';
-import { serverHandlers } from './handlers.server';
+import { serverHandlers } from './handlers';
 
 export const server = setupServer(...serverHandlers);
